Html table working
Basically, i've gotten the file to upload as an html table using the 'react-csv-to-table' import. After this i'm going to be implementing material design, as well as implementing removing of entries, as well as updating of entries, and lastly  refactoring how entries are added. After all of that is done, i'll fully transfer over to material design, clean up the code, and possibly turn this into an api that others can easily use instead of having to go through all the trouble.